feat: add CampaignDetailPage and CampaignForm components for campaign management; implement form handling and validation
